Update and rename OutfitManager.js to BotOutfitManager.js
