+++ src/BotOutfitManager.js
@@ -1,0 +1,83 @@
+import { extension_settings } from "../../../../extensions.js";
+
+export class BotOutfitManager {
+    constructor(slots) {
+        this.slots = slots;
+        this.character = 'Unknown';
+        this.currentValues = {};
+        this.slots.forEach(slot => this.currentValues[slot] = 'None');
+    }
+
+    setCharacter(name) {
+        if (name === this.character) return;
+        this.character = name;
+        this.loadOutfit();
+    }
+
+    getVarName(slot) {
+        return `${this.character.replace(/\s+/g, '_')}_${slot}`;
+    }
+
+    loadOutfit() {
+        this.slots.forEach(slot => {
+            const varName = this.getVarName(slot);
+            this.currentValues[slot] = this.getGlobalVariable(varName) || 'None';
+        });
+    }
+
+    getGlobalVariable(name) {
+        return window[name] || (extension_settings.variables?.global?.[name] || 'None');
+    }
+
+    setGlobalVariable(name, value) {
+        window[name] = value;
+        if (!extension_settings.variables) extension_settings.variables = { global: {} };
+        extension_settings.variables.global[name] = value;
+    }
+
+    async setOutfitItem(slot, value) {
+        const previousValue = this.currentValues[slot];
+        const varName = this.getVarName(slot);
+        this.setGlobalVariable(varName, value);
+        this.currentValues[slot] = value;
+
+        if (previousValue === 'None' && value !== 'None') {
+            return `${this.character} put on ${value}.`;
+        } else if (value === 'None') {
+            return `${this.character} removed ${previousValue}.`;
+        } else {
+            return `${this.character} changed from ${previousValue} to ${value}.`;
+        }
+    }
+
+    async changeOutfitItem(slot) {
+        const currentValue = this.currentValues[slot];
+        let newValue = currentValue;
+
+        if (currentValue === 'None') {
+            newValue = prompt(`What is ${this.character} wearing on their ${slot}?`, "");
+            if (!newValue) return null;
+        } else {
+            const choice = prompt(
+                `${this.character}'s ${slot}: ${currentValue}\n\nEnter 'remove' to remove, or type new item:`,
+                ""
+            );
+
+            if (!choice) return null;
+            newValue = choice.toLowerCase() === 'remove' ? 'None' : choice;
+        }
+
+        if (newValue !== currentValue) {
+            return this.setOutfitItem(slot, newValue);
+        }
+        return null;
+    }
+
+    getOutfitData() {
+        return this.slots.map(slot => ({
+            name: slot,
+            value: this.currentValues[slot],
+            varName: this.getVarName(slot)
+        }));
+    }
+}
